use uint32_t textureId

+++ Src/Core/Resources/Headers/Model.hpp
@@ -24,7 +24,7 @@ namespace AnA
             glm::vec3 position{};
             uint16_t pitch{};
             uint16_t yaw{};
-            uint16_t textureId;
+            uint32_t textureId;
             glm::u8vec3 color{};
             glm::vec2 uv{};
